Update 0.1.1
minor changes to script
