+++ Assets/Michael's Folder/PlayerMechanics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 733d0bab54cbf1c4f909ecf7574f073b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
