+++ Assets/Michael's Folder/PlayerMechanics/RockPickUp.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b555af023caf2144c836ada8158b209c
+guid: b8486495b8b218c48b8d57422df1514b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
