--- Assets/Michael's Folder/RockSkip.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d0b0972afb701384d985c2c17381ac61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
