--- Assets/Michael's Folder/Rockmechanics/RockThrow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e580fe76aa209df4ca2bbc3395adc17b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
