+++ Assets/Michael's Folder/Rockmechanics/RockThrowing.cs
@@ -2,33 +2,33 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class RockThrow : MonoBehaviour
+public class RockThrowing : MonoBehaviour
 {
     public GameObject rock;
-    
-    float launchSpeed
+
+    float launchSpeed;
     
 
     // Start is called before the first frame update
     void Start()
     {
-       RockThrow = 
+       rock = GetComponent<GameObject>();
     }
 
     // Update is called once per frame
     void Update()
     {
         //throw when button is released
-        Input.GetButtonUp(One){
-            Throw();
+        if(Input.GetButtonUp("One"))
+        {
+            Throw(rock);
         }
     }
 
-    public void Throw(GameObject rock)
+    public void Throw(GameObject rock) //lauch object by instantiating it with a given speed
     {
         //instantiate object with launchSpeed and direction of player
         Instantiate(rock); 
     } 
 
-    public void 
 }
